+++ NEWS.md
@@ -1,5 +1,12 @@
 # fs (development version)
 
+* New `path_select_components()` function to select components of one or
+  more paths (#326, @Tazinho).
+
+# fs 1.6.6
+
+* No changes.
+
 # fs 1.6.5
 
 * `path_ext()` and `path_ext_remove()` return correct extension and path, respectively, when multiple dots are present in file name (@IndrajeetPatil, #452, #453).